shell.nix: Add CL-VULKAN dependency (and SBCL init file)
Add CL-VULKAN as a dependency.

Also add an initialization file to SBCL to control settings
like *FEATURES* when dependencies are compiled. Forgive me, this is
currently used for tweaking CL-OPENGL settings, which are not relevant
to the branch this commit is on... so this part of the commit is just
following along for the ride at the moment.