(fix) ray: use appropriate function to instantiate ray
This commit fixes some unexpected behaviour of object picking. See the
note (`observation`) note towards the end of this comment:
https://github.com/kaveh808/kons-9/pull/222

The problem was with the function being used for ray creation.  The origin
library provides a couple of functions to create a ray. Prior to this, we used a
helper function which expected the ray `origin` and `direction` as parameters
while what was actually being passed were the `from` and `to` coordinates. In
this commit we use the more appropriate function which expects the `from` and
`to` coordinates.